[폴더위계정리] refactor: SNUTTUtils.kt dead code 제거 및 분해
미사용 함수 9개(wdayToNumber, numberToWday, numberToTime 등) 삭제 후
남은 함수를 유일 사용처로 이동하고 파일 삭제:
- semesterStringToLong → DeeplinkParser.kt private 함수
- displayWidth/displayHeight → ViewExt.kt
